requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
argon2-cffi
python-jose[cryptography]
pydantic[email]
